chore: move common functions to shared package
