+++ internal/storage/ls/ls.go
@@ -2,14 +2,13 @@ package ls
 
 import (
 	"context"
-	"errors"
 	"fmt"
-	"net/url"
 	"os"
 	"path"
 	"strings"
 
 	"github.com/spf13/afero"
+	"github.com/supabase/cli/internal/storage"
 	"github.com/supabase/cli/internal/storage/client"
 	"github.com/supabase/cli/internal/utils"
 )
@@ -17,7 +16,7 @@ import (
 const STORAGE_SCHEME = "ss"
 
 func Run(ctx context.Context, objectPath string, recursive bool, fsys afero.Fs) error {
-	remotePath, err := ParseStorageURL(objectPath)
+	remotePath, err := storage.ParseStorageURL(objectPath)
 	if err != nil {
 		return err
 	}
@@ -35,19 +34,6 @@ func Run(ctx context.Context, objectPath string, recursive bool, fsys afero.Fs) 
 	return IterateStoragePaths(ctx, projectRef, remotePath, callback)
 }
 
-var errInvalidURL = errors.New("URL must match pattern ss:///bucket/[prefix]")
-
-func ParseStorageURL(objectPath string) (string, error) {
-	parsed, err := url.Parse(objectPath)
-	if err != nil {
-		return "", err
-	}
-	if strings.ToLower(parsed.Scheme) != STORAGE_SCHEME || len(parsed.Path) == 0 || len(parsed.Host) > 0 {
-		return "", errInvalidURL
-	}
-	return parsed.Path, nil
-}
-
 func ListStoragePaths(ctx context.Context, projectRef, remotePath string) ([]string, error) {
 	var result []string
 	err := IterateStoragePaths(ctx, projectRef, remotePath, func(objectName string) error {
@@ -58,7 +44,7 @@ func ListStoragePaths(ctx context.Context, projectRef, remotePath string) ([]str
 }
 
 func IterateStoragePaths(ctx context.Context, projectRef, remotePath string, callback func(objectName string) error) error {
-	bucket, prefix := SplitBucketPrefix(remotePath)
+	bucket, prefix := storage.SplitBucketPrefix(remotePath)
 	if len(bucket) == 0 || (len(prefix) == 0 && !strings.HasSuffix(remotePath, "/")) {
 		buckets, err := client.ListStorageBuckets(ctx, projectRef)
 		if err != nil {
@@ -95,21 +81,6 @@ func IterateStoragePaths(ctx context.Context, projectRef, remotePath string, cal
 		}
 	}
 	return nil
-}
-
-func SplitBucketPrefix(objectPath string) (string, string) {
-	if objectPath == "" || objectPath == "/" {
-		return "", ""
-	}
-	start := 0
-	if objectPath[0] == '/' {
-		start = 1
-	}
-	sep := strings.IndexByte(objectPath[start:], '/')
-	if sep < 0 {
-		return objectPath[start:], ""
-	}
-	return objectPath[start : sep+start], objectPath[sep+start+1:]
 }
 
 // Expects remotePath to be terminated by "/"
@@ -156,7 +127,7 @@ func IterateStoragePathsAll(ctx context.Context, projectRef, remotePath string, 
 			return err
 		}
 		// Also report empty buckets
-		bucket, prefix := SplitBucketPrefix(dirPath)
+		bucket, prefix := storage.SplitBucketPrefix(dirPath)
 		if empty && len(prefix) == 0 {
 			if err := callback(bucket + "/"); err != nil {
 				return err
